Fixed meta range

+++ CrayEditAPI/src/main/java/jp/llv/ce/server/EditMaterialData.java
@@ -33,7 +33,7 @@ public interface EditMaterialData {
     
     EditMaterial getMaterial();
     
-    byte getData();
+    short getData();
     
     Map getTag();
